+++ adapters/Pangle/PangleAdapter/GADMediationAdapterPangle.m
@@ -123,9 +123,7 @@ static NSInteger _GDPRConsent = -1, _doNotSell = -1;
 
 - (void)loadBannerForAdConfiguration:(GADMediationBannerAdConfiguration *)adConfiguration
                    completionHandler:(GADMediationBannerLoadCompletionHandler)completionHandler {
-  [GADMediationAdapterPangle setCOPPA:(adConfiguration.childDirectedTreatment
-                                           ? adConfiguration.childDirectedTreatment.integerValue
-                                           : -1)];
+  [GADMediationAdapterPangle setCOPPA];
   _bannerRenderer = [[GADPangleBannerRenderer alloc] init];
   [_bannerRenderer renderBannerForAdConfiguration:adConfiguration
                                 completionHandler:completionHandler];
@@ -135,9 +133,7 @@ static NSInteger _GDPRConsent = -1, _doNotSell = -1;
             (GADMediationInterstitialAdConfiguration *)adConfiguration
                          completionHandler:
                              (GADMediationInterstitialLoadCompletionHandler)completionHandler {
-  [GADMediationAdapterPangle setCOPPA:(adConfiguration.childDirectedTreatment
-                                           ? adConfiguration.childDirectedTreatment.integerValue
-                                           : -1)];
+  [GADMediationAdapterPangle setCOPPA];
   _interstitialRenderer = [[GADPangleInterstitialRenderer alloc] init];
   [_interstitialRenderer renderInterstitialForAdConfiguration:adConfiguration
                                             completionHandler:completionHandler];
@@ -146,9 +142,7 @@ static NSInteger _GDPRConsent = -1, _doNotSell = -1;
 - (void)loadRewardedAdForAdConfiguration:(GADMediationRewardedAdConfiguration *)adConfiguration
                        completionHandler:
                            (GADMediationRewardedLoadCompletionHandler)completionHandler {
-  [GADMediationAdapterPangle setCOPPA:(adConfiguration.childDirectedTreatment
-                                           ? adConfiguration.childDirectedTreatment.integerValue
-                                           : -1)];
+  [GADMediationAdapterPangle setCOPPA];
   _rewardedRenderer = [[GADPangleRewardedRenderer alloc] init];
   [_rewardedRenderer renderRewardedAdForAdConfiguration:adConfiguration
                                       completionHandler:completionHandler];
@@ -157,9 +151,7 @@ static NSInteger _GDPRConsent = -1, _doNotSell = -1;
 - (void)loadNativeAdForAdConfiguration:(nonnull GADMediationNativeAdConfiguration *)adConfiguration
                      completionHandler:
                          (nonnull GADMediationNativeLoadCompletionHandler)completionHandler {
-  [GADMediationAdapterPangle setCOPPA:(adConfiguration.childDirectedTreatment
-                                           ? adConfiguration.childDirectedTreatment.integerValue
-                                           : -1)];
+  [GADMediationAdapterPangle setCOPPA];
   _nativeRenderer = [[GADPangleNativeRenderer alloc] init];
   [_nativeRenderer renderNativeAdForAdConfiguration:adConfiguration
                                   completionHandler:completionHandler];
@@ -169,26 +161,22 @@ static NSInteger _GDPRConsent = -1, _doNotSell = -1;
             (nonnull GADMediationAppOpenAdConfiguration *)adConfiguration
                       completionHandler:
                           (nonnull GADMediationAppOpenLoadCompletionHandler)completionHandler {
-  [GADMediationAdapterPangle setCOPPA:(adConfiguration.childDirectedTreatment
-                                           ? adConfiguration.childDirectedTreatment.integerValue
-                                           : -1)];
+  [GADMediationAdapterPangle setCOPPA];
   _appOpenAdRenderer = [[GADPangleAppOpenRenderer alloc] init];
   [_appOpenAdRenderer renderAppOpenAdForAdConfiguration:adConfiguration
                                       completionHandler:completionHandler];
 }
 
 /// Set the COPPA setting in Pangle SDK.
-///
-/// @param COPPA An integer value that indicates whether the app should be treated as
-/// child-directed for purposes of the COPPA.  0 means false. 1 means true. -1 means
-/// unspecified. Any value outside of -1, 0, or 1 will result in this method being a no-op.
-+ (void)setCOPPA:(NSInteger)COPPA {
-  if (COPPA != 0 && COPPA != 1 && COPPA != -1) {
-    GADMPangleLog(@"Invalid COPPA value. Pangle SDK only accepts -1, 0 or 1.");
-    return;
-  }
-  if (PAGSdk.initializationState == PAGSDKInitializationStateReady) {
-    PAGConfig.shareConfig.childDirected = COPPA;
++ (void)setCOPPA {
+  NSNumber *tagForChildDirectedTreatment =
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment;
+  if (!tagForChildDirectedTreatment) {
+    PAGConfig.shareConfig.childDirected = PAGChildDirectedTypeDefault;
+  } else if ([tagForChildDirectedTreatment boolValue] == YES) {
+    PAGConfig.shareConfig.childDirected = PAGChildDirectedTypeChild;
+  } else if ([tagForChildDirectedTreatment boolValue] == NO) {
+    PAGConfig.shareConfig.childDirected = PAGChildDirectedTypeNonChild;
   }
 }