+++ adapters/Pangle/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Pangle iOS Mediation Adapter Changelog
 
+#### Next version
+- Refactor setCOPPA by replacing deprecated adConfiguration.childDirectedTreatment with new GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment.
+
 #### [Version 5.5.0.9.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-5.5.0.9.0.zip)
 - Verified compatibility with Pangle SDK version 5.5.0.9.